% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deduplicate_dqa.R
\name{deduplicate_dqa}
\alias{deduplicate_dqa}
\title{Deduplicate DQA (Data Quality Assessment) data.}
\usage{
deduplicate_dqa(df)
}
\arguments{
\item{df}{A dataframe containing DQA data.}
}
\value{
A deduplicated dataframe.
}
\description{
This function identifies and removes duplicate records in the DQA data based on specified criteria.
}
\examples{
# Create sample dataframe
df <- data.frame(
  facility = c("A", "A", "B", "B", "C", "C"),
  client_id = c(1, 2, 1, 3, 4, 2),
  abstracted_days_of_arv_refill = c(30, 28, 30, 28, 30, 28),
  abstracted_art_start_date = as.Date(c("2022-01-01", "2022-02-01", "2022-03-01", "2022-04-01", NA, NA)),
  abstracted_last_pickup_date = as.Date(c("2022-01-28", "2022-02-28", "2022-03-28", "2022-04-28", "2022-05-28", "2022-06-28"))
)

# Deduplicate DQA data
result <- deduplicate_dqa(df)

}
