% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_non_concurrence.R
\name{get_non_concurrence}
\alias{get_non_concurrence}
\title{Get non-concurring records between cleaned and extracted data.}
\usage{
get_non_concurrence(df)
}
\arguments{
\item{df}{A dataframe containing cleaned and extracted data.}
}
\value{
A dataframe with non-concurring records.
}
\description{
This function identifies records that do not match between cleaned and extracted data based on specified criteria.
}
\examples{
# Create sample dataframe
df <- data.frame(
  facility = c("A", "A", "B", "B", "C", "C"),
  sex = c("M", "F", "M", "F", "M", "F"),
  art_start_date = as.Date(c("2022-01-01", "2022-02-01", "2022-03-01", "2022-04-01", NA, NA)),
  last_pickup_date = as.Date(c("2022-01-28", "2022-02-28", "2022-03-28", "2022-04-28", "2022-05-28", "2022-06-28")),
  days_of_arv_refill = c(30, 28, 30, 28, 30, 28),
  abstracted_sex = c("M", "F", "M", "F", "M", "F"),
  abstracted_art_start_date = as.Date(c("2022-01-01", "2022-02-01", "2022-03-01", "2022-04-01", "2022-05-01", "2022-06-01")),
  abstracted_last_pickup_date = as.Date(c("2022-01-28", "2022-02-28", "2022-03-28", "2022-04-28", "2022-05-28", "2022-06-28")),
  abstracted_days_of_arv_refill = c(30, 28, 30, 28, 30, 28)
)

# Get non-concurring records
result <- get_non_concurrence(df)

}
